all_together/gjs-pc-requirements.txt
backoff==1.8.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
ciso8601==2.2.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
jsonobject==2.0.0
jsonschema==2.6.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
pytz==2018.4
PyYAML==6.0
requests==2.24.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.11.1
singer-python==5.9.0
six==1.16.0
SQLAlchemy==1.4.39
tap-ujet==1.0.0
urllib3==1.25.11
Werkzeug==2.2.1

